pyproject.toml
[project]
name = "call-center-coach"
version = "0.1.0"
description = "A demo about how AI can be a Sales Coach"
readme = "README.md"
requires-python = ">=3.11, <3.12"
dependencies = [
    "ffmpeg-python>=0.2.0",
    "openai-whisper>=20240930",
    "openai>=1.66.3",
    "pydub>=0.25.1",
    "streamlit>=1.43.2",
    "torch>=2.6.0",
    "torchvision>=0.21.0",
    "torchaudio>=2.6.0",
    "watchdog>=6.0.0",
    "python-dotenv>=1.0.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.0",
]
